requirements.txt
# Django and DRF
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter
# mysql-connector-python>=8.0.0  # Uncomment for MySQL

# CORS handling
django-cors-headers>=4.3.0

# Environment variables
python-decouple>=3.8
python-dotenv>=1.0.0

# Password validation and security
argon2-cffi>=23.1.0  # Better password hashing
cryptography>=41.0.0

# API documentation
drf-spectacular>=0.26.0

# Development and testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0
coverage>=7.3.0

# Code quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.5.0  # Static file serving

# Monitoring and logging
sentry-sdk>=1.32.0

# Utilities
Pillow>=10.0.0  # Image processing
celery>=5.3.0  # Task queue (optional)
redis>=4.6.0  # Caching and message broker

# Additional dependencies from master branch
dj-database-url==2.1.0
requests==2.32.4 
